Fix Deployment error caused by Organization Commit

+++ Class4910Api/Class4910Api/Services/Interfaces/IOrganizationService.cs
@@ -11,7 +11,7 @@ public interface IOrganizationService
 
     Task<Organization?> GetOrganizationByName(string orgName);
 
-    Task<Organization?> UpdateOrganizationPointValue(int organizationId, float newPointValue, int updaterUserId);
+    Task<Organization?> UpdateOrganizationPointValue(int organizationId, double newPointValue, int updaterUserId);
 
     Task<List<Organization>?> GetOrganizations();
 }
